[#26] move(Post): EducationData 패키지명 Post로 변경
