+++ src/main/java/kr/sparta/rchive/domain/bookmark/entity/Bookmark.java
@@ -6,8 +6,7 @@ import jakarta.persistence.IdClass;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
-import kr.sparta.rchive.domain.educationData.entity.EducationData;
-import kr.sparta.rchive.domain.educationData.entity.EducationDataTrackId;
+import kr.sparta.rchive.domain.post.entity.EducationData;
 import kr.sparta.rchive.domain.user.entity.User;
 import kr.sparta.rchive.global.entity.BaseTimeEntity;
 import lombok.AllArgsConstructor;
