+++ src/main/java/kr/sparta/rchive/domain/comment/entity/Comment.java
@@ -2,8 +2,6 @@ package kr.sparta.rchive.domain.comment.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
@@ -11,10 +9,8 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import java.time.LocalDateTime;
-import kr.sparta.rchive.domain.educationData.entity.EducationData;
-import kr.sparta.rchive.domain.user.entity.Track;
+import kr.sparta.rchive.domain.post.entity.EducationData;
 import kr.sparta.rchive.domain.user.entity.User;
-import kr.sparta.rchive.domain.user.enums.OauthTypeEnum;
 import kr.sparta.rchive.global.entity.BaseTimeEntity;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
