+++ src/main/java/kr/sparta/rchive/domain/core/service/CoreService.java
@@ -1,5 +1,5 @@
 package kr.sparta.rchive.domain.core.service;
 
-public class coreservice {
+public class CoreService {
 
 }
