+++ src/main/java/kr/sparta/rchive/domain/post/entity/EducationDataTrack.java
@@ -1,4 +1,4 @@
-package kr.sparta.rchive.domain.educationData.entity;
+package kr.sparta.rchive.domain.post.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
@@ -6,6 +6,7 @@ import jakarta.persistence.IdClass;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
+import kr.sparta.rchive.domain.user.entity.Track;
 import kr.sparta.rchive.global.entity.BaseTimeEntity;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -18,8 +19,9 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-@IdClass(EducationDataTagId.class)
-public class EducationDataTag extends BaseTimeEntity {
+@IdClass(EducationDataTrackId.class)
+
+public class EducationDataTrack extends BaseTimeEntity {
 
     @Id
     @ManyToOne
@@ -28,6 +30,7 @@ public class EducationDataTag extends BaseTimeEntity {
 
     @Id
     @ManyToOne
-    @JoinColumn(name = "tag_id")
-    private Tag tag;
+    @JoinColumn(name = "track_id")
+    private Track track;
+
 }
