+++ src/main/java/kr/sparta/rchive/domain/post/entity/Tag.java
@@ -1,4 +1,4 @@
-package kr.sparta.rchive.domain.educationData.entity;
+package kr.sparta.rchive.domain.post.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
