+++ src/main/java/kr/sparta/rchive/domain/post/enums/EducationTypeEnum.java
@@ -1,4 +1,4 @@
-package kr.sparta.rchive.domain.educationData.enums;
+package kr.sparta.rchive.domain.post.enums;
 
 public enum EducationTypeEnum {
     Sparta_Lecture,
